server: Move RestApiTestConfiguration to a dedicated file
